requirements.txt
absl-py==2.3.1
amqp==5.3.1
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
autobahn==24.4.2
automat==25.4.16
billiard==4.2.2
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels-redis==4.2.1
charset-normalizer==3.4.2
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
contourpy==1.3.3
cryptography==45.0.2
cssselect2==0.8.0
cycler==0.12.1
daphne==4.2.0
decorator==5.2.1
django==5.2.1
django-allauth==65.8.1
django-cors-headers==4.7.0
django-easy-audit==1.3.7
django-environ==0.12.0
django-filter==25.1
django-htmx==1.23.0
django-tailwind==4.0.1
django-unfold==0.58.0
djangorestframework==3.16.0
drf-yasg==1.21.10
executing==2.2.0
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
freetype-py==2.5.1
fsspec==2025.9.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
ipython==9.2.0
ipython-pygments-lexers==1.1.1
jax==0.8.0
jaxlib==0.8.0
jedi==0.19.2
jinja2==3.1.6
kiwisolver==1.4.9
kombu==5.5.4
loguru==0.7.3
lxml==6.0.2
markdown==3.8
markupsafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
mediapipe==0.10.14
ml-dtypes==0.5.3
mpmath==1.3.0
msgpack==1.1.2
networkx==3.5
numpy==2.2.6
oauthlib==3.2.2
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt-einsum==3.4.0
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pillow==12.0.0
polars==1.34.0
polars-runtime-32==1.34.0
prompt-toolkit==3.0.51
protobuf==4.25.8
psutil==7.1.2
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycairo==1.28.0
pycparser==2.22
pygments==2.19.1
pyjwt==2.10.1
pyopenssl==25.1.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
redis==5.2.1
reportlab==4.4.4
requests==2.32.3
requests-oauthlib==2.0.0
rlpycairo==0.4.0
scipy==1.16.3
service-identity==24.2.0
setuptools==80.8.0
six==1.17.0
sounddevice==0.5.3
sqlparse==0.5.3
stack-data==0.6.3
svglib==1.6.0
sympy==1.14.0
tinycss2==1.4.0
torch==2.9.0
torchvision==0.24.0
traitlets==5.14.3
twisted==24.11.0
txaio==23.1.1
typing-extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.222
ultralytics-thop==2.0.18
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
whitenoise==6.9.0
zope-interface==7.2
